+++ include/VideoCaptureFactory.hpp
@@ -1,11 +1,6 @@
 #pragma once
-#include "VideoCaptureInterface.hpp"
-#ifdef USE_GSTREAMER
-#include "GStreamerCapture.hpp"
-#endif
-#ifdef USE_FFMPEG
-#include "FFmpegCapture.hpp"
-#endif
-#include "OpenCVCapture.hpp"
 
- std::unique_ptr<VideoCaptureInterface> createVideoInterface(); 
+#include <memory>
+#include "VideoCaptureInterface.hpp"
+
+std::unique_ptr<VideoCaptureInterface> createVideoInterface();