+++ src/opencv/OpenCVCapture.hpp
@@ -1,17 +1,17 @@
 #pragma once
-#include "VideoCaptureInterface.hpp"
-#include <opencv2/highgui/highgui.hpp>
 
+#include <opencv2/videoio.hpp>
+#include "VideoCaptureInterface.hpp"
 
 class OpenCVCapture : public VideoCaptureInterface {
 private:
     cv::VideoCapture capture;
-    bool initialized = false; // Track initialization status
+    bool initialized = false;  // Track initialization status
 
 public:
     bool initialize(const std::string& source) override;
 
-    bool readFrame(cv::Mat& frame) override;
+    bool readFrame(VideoFrame& frame) override;
 
     void release() override;
 };